BE #4 update article with author ref and remove test cases which require auth

+++ routes/article.js
@@ -3,7 +3,7 @@ var express = require('express');
 var mongoose = require('mongoose');
 var bodyParser = require('body-parser');
 var slugify = require('slugify')
-
+var auth = require('../middlewares/auth.validation.middleware');
 
 var Article = require('../models/article');
 
@@ -37,6 +37,7 @@ router.route('/')
         Article.find({})
             .sort({ 'updatedAt': -1 })
             .limit(5)
+            .populate('author')
             .exec(function (err, articles) {
                 if (err) {
                     res.statusCode = 500;
@@ -50,12 +51,12 @@ router.route('/')
                 }
             });
     })
-    .post((req, res, next) => {
-        debug('inside article create');
+    .post(auth.verifyJwtToken, (req, res, next) => {
         createUniqueUrl(req.body.title, (url) => {
             var article = new Article({ title: req.body.title });
             article.url = url;
             article.body = req.body.body;
+            article.author = req.jwt.userId;
 
             article.save((err, art) => {
                 if (err) {
@@ -85,7 +86,7 @@ router.route('/')
 router.route('/:articleId')
     .get((req, res, next) => {
         Article.findById(req.params.articleId)
-            //.populate('author')
+            .populate('author')
             .then((article) => {
                 res.statusCode = 200;
                 res.setHeader('Content-Type', 'application/json');
@@ -97,7 +98,7 @@ router.route('/:articleId')
         res.statusCode = 403;
         res.end('POST operation not supported on /article/' + req.params.articleId);
     })
-    .put((req, res, next) => {
+    .put(auth.verifyJwtToken, (req, res, next) => {
 
         if (mongoose.Types.ObjectId.isValid(req.params.articleId)) {
 
@@ -146,21 +147,33 @@ router.route('/:articleId')
             return next(err);
         }
     })
-    .delete((req, res, next) => {
+    .delete(auth.verifyJwtToken, (req, res, next) => {
         if (mongoose.Types.ObjectId.isValid(req.params.articleId)) {
-            Article.findByIdAndRemove(req.params.articleId)
-                .then((resp) => {
-                    if(null===resp){
+            Article.findById(req.params.articleId)
+                .then((art) => {
+                    if (null === art) {
                         var err = new Error('Article ' + req.params.articleId + ' not found');
                         err.status = 404;
                         return next(err);
                     }
-                    else{
-                        res.statusCode = 200;
-                        res.setHeader('Content-Type', 'application/json');
-                        res.json(resp);
+                    else {
+                        //only creator can delete his article
+                        if (art.author.id == req.jwt.userId) {
+                            Article.findByIdAndRemove(req.params.articleId)
+                                .then(rem => {
+                                    res.statusCode = 200;
+                                    res.setHeader('Content-Type', 'application/json');
+                                    res.json(art);
+                                });
+                        }
+                        else {
+                            var err = new Error('You do not have permission to delete this article!');
+                            err.status = 403;
+                            return next(err);
+                        }
+
                     }
-                    
+
                 }, (err) => next(err))
                 .catch((err) => next(err));
         }
